+++ src/main/resources/applicationContext.xml
@@ -13,7 +13,7 @@
         <property name="password" value="sa"/>
     </bean>
 
-    <bean id="userDao" class="springbook.user.dao.UserDao">
+    <bean id="userDao" class="springbook.user.dao.UserDaoJdbc">
         <property name="dataSource" ref="dataSource"/>
     </bean>
